Fix for handling unit dates

+++ src/Build.php
@@ -132,11 +132,11 @@ class Build extends Arc\Singleton
 					 `gmr`                INT(9) DEFAULT 0,
 					 `netAskingRental`    INT(9) DEFAULT 0,
 					 `availableType`      VARCHAR(90),
-					 `availableFrom`      FLOAT(11,4) DEFAULT 0,
+					 `availableFrom`      INT(11) DEFAULT 0,
 					 `marketingHeading`   VARCHAR(90),
 					 `description`        BLOB,
 					 `updated_at`         INT(11) NOT NULL,
-					 `gmv_updated`        FLOAT(11,4) NOT NULL
+					 `gmv_updated`        INT(11) NOT NULL
 					)
 				";
 				$db->exec($q);
